[fea-rs] Move DiagnosticSet into diagnostics module
And related cleanup.
