+++ fea-rs/src/lib.rs
@@ -17,6 +17,6 @@ mod tests;
 
 pub use common::{GlyphIdent, GlyphMap, GlyphName, GlyphSet};
 pub use compile::Compiler;
-pub use diagnostic::{Diagnostic, Level};
+pub use diagnostic::{Diagnostic, DiagnosticSet, Level};
 pub use parse::{ParseTree, TokenSet};
 pub use token_tree::{typed, Kind, Node, NodeOrToken, Token};
